+++ NEWS.md
@@ -1,4 +1,10 @@
 
+## fuzzywuzzyR 1.0.2
+
+I added the *encoding* parameter to the following classes : *FuzzExtract*, *FuzzMatcher* and *FuzzUtils*. The *decoding* parameter does not apply to the *GetCloseMatches* and *SequenceMatcher* classes, because there isn't any *force_ascii* parameter in the *difflib* python library.
+For reference, see the following github issue : https://github.com/mlampros/fuzzywuzzyR/issues/3
+
+
 ## fuzzywuzzyR 1.0.1
 
 I added links to the github repository (master repository, issues).